Use the correct SHA for pull request snapshots
Evidently, `github.context.sha` is not always what you might consider
the "current commit" in a given context. In most GitHub pull request
event types, `github.context.sha` will be the "Last merge commit on the
GITHUB_REF branch." In those cases, the commit SHA that should be
associated with the snapshot is `github.event.pull_request.head.sha`.

This commit adds a helper function, `getSha`, which will return the
correct SHA for the current context.

Signed-off-by: Justin Holguin <juxtin@github.com>
Signed-off-by: GitHub <noreply@github.com>